Implementing proc
devproc needs to be able to get the first, second, ... processes in
the hash table. To get this, implement pid_nth. Possibly, we need a
general function this like this in hash.c, not a special-purpose one
in process.c

cat '#p' now works correctly.
ls '#p' does not, and the failure is weird; kernel locks up. G works not.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>