Simple root file system driver.
It's #r and has net, proc, and bin. This is arguably a stupid
way to do it but it is kind of a nice tutorial, it took me about 10
minutes to cons it up and test it.

Once we have a bind user-level binary (system calls are done)
we can do stuff to test like
bind '#p' '#r/proc'
and then
ls '#r/proc'
and see processes. Once we get to
bind '#r' /

then we have root.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>