Testing 9 name space.
This includes some cleanup and fixes and a nasty patch to make testing
easier.

The cleanups are obvious.

The nasty patch is this: in sysopen and sysstat, if the first two
chars are /9, they are skipped and the rest of the path resolution
processed with the remaining path.

It's proving to be a very easy way to test out the name space. I agree
it's ugly.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>