Cleans up 9ns init for new processes
Doesn't handle any of the plan9 rfork flags yet.

Note the kref acrobatics.  A concurrent thread could be affecting the parent's
dot (though not the slash), so we need to be careful about getting our ref.  To
trigger it, you'd need one thread calling cd while another created a process,
and to be lucky.