add in plan 9 command parsing
Used in drivers. Don't get too upset when you see parsing in the kernel.
It has to be done somewhere, and when you do it this way, you can
mount a 386 device on and arm and issue commands to it without
worrying about endianness, word size, and all that stuff.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>