Improve devrandom and the test
I've verified that the random device will make reasonable attempts
to use system activity to create random numbers that are less non-random.

The trandom test now takes an arg. It defaults to 4000 bytes. If there
is not much activity on your system it will take a long time to run.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>