Adds a pcpu tracer for spinlocks
Compile with SPINLOCK_DEBUG and TRACE_LOCKS, then access the info from the
monitor with 'trace pcpui'.

This is very spammy, and it'll clobber any other info in the trace, so be
careful using it.  It will also attempt to find the location of the lock, based
on the symtab.  This only works for locks above KLA (o/w, the lock is dynamic).

When building with TRACE_LOCKS, we also disable any of the fancier checks for
the console and printf locks, so that they don't clobber our results.
Otherwise, as you printed lines, you'd be adding to the trace.