Get dup to work correctly.
busybox depends heavily on dup and I got it wrong first time around.

e.g. dd if=xyz
will open zyx, dup it onto 0, and read from 0.

The plan9fd has to be in the fd array, and the plan 9 sentinal
has to be in the file struct.

This now seems to work as it should.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>