Get proc working.
The issue is that in plan 9 the proc table is an array, and in
akaros it's a hash. We should change dev/proc.c to always use
the hash but this will require redoing how the proc QID works
and there are bigger fish to fry.

Next step is to get '#p/1/ns' to work. Then we can see our
name spaces.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>