Random device now works for atomic pipes
The testrandom now prints the last four u32ints read. This
does not change the timing result.

The original random device
does a ^ operation with the data from 1024 generations ago
to produce the new data. To simulate this we keep an independent
counter as an index into the array -- we go behind the back
of the atomic pipes. Is this a good idea? Time will tell.

Statistics:
  89  212 1606 kern/drivers/dev/newrandom.c
 166  382 2969 kern/drivers/dev/random.c

A significant reduction in complexity,
a significant improvement in
readability.
The numbers generated suck as random numbers so
that needs to be fixed.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>