Cleans up plan9.h (XCC)
Some of the #defines are useful for the entire kernel.  Those were moved
to ros/common.h (which is actually part of the kernel headers, so
userspace can use them too).

ARRAY_SIZE() was defined twice, once as ARRAY_SIZE, and later as nelms.

Some of the bits for syscalls were listed twice too.

Reinstall your kernel headers to get access to the new #defines from
userspace.