Set errno when only errstr was called
Some error paths set errstr, but not errno.  Glibc will ignore errstr if there
isn't an errno too, so we need something.

Ultimately, we should be setting both errno and errstr (given a glibc-world
with errno).  Til then, we'll set a generic errno.

This won't harm the current glibc, which 0's sysc.errstr before calling, but
that isn't actually part of the ABI.  So with this hack, it is possible to set
errno when there wasn't actually an error.  Luckily, userspace shouldn't look
at errno unless it was told too (via a bad retval), in which case, we probably
wanted errno set anyway.  Still, be careful.