Clean up ping and get it to at least try to send a packet
And, as expected, our atomic pipe usage is not quite right:
Entering Nanwan's Dungeon on Core 0 (Ints on):
Type 'help' for a list of commands.
ROS(Core 0)> backtrace
Stack Backtrace on Core 0:

Note that it's a good idea to use 127.1 for now; this will test
the medium layer before we go to try to test a real interface.
We can chase out bugs in qio code.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>