Make set_errstr take a printf-like format
One of the things I really disliked about Plan 9 was that we
could return an errstr, and we only ever returned 52 errors
even though errstr was called in over 112 places. It gets
even more confusing if a program or library transforms the args:
what did the args look like precisely? Hard to say.

So on Plan 9 we get errors no more informative than errno. That's a very
low bar, to say the least.

This change allows for more useful error messages. We include an
example -- not a great one, but hey, it already cleared something
up for me.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>