Fixes apipe_read_cond()
Uses a priority_reader CV, so that we can have only one reader at a
time.  In the process, I separated the reader and writer CVs too.  Since
we have multiple CVs sharing a lock, it's clearer to use a standalone
lock.  Maybe I'll switch over to the POSIX CV style in the future...

Note: this was cherry-picked and had some minor issues.  I blame any bugs on
that.