Ping changes
There's no shared memory fork; ping must run serially.

Weirdly, we seem to always read the same packet back. Or there are different
gcc things going on than kenc. Not sure yet.

But the seq in the receive packet is always the same value. But in write, bug
in read, but in qio? Who knows?

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>