extend devroot.c to support /bin and /lib
This is in preparation to making it a full ramfs. Next will be
file creation and writing. Then directories. Crawl, walk, fly.

Also, feel free to finish this, I'm going to be absent for a day or two.
It's a pretty trivial set of changes.

There are two static files in /bin and /lib and reading them works.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>