Kthreads track cur_sysc and cur_errbuf
Instead of having them sit in pcpui, then copying them to the kthread,
we just put them directly in the kthread.  This matches their intent.
Using pcpui was more of a hack back when we didn't have kthreads at all
times.

Note that set_errno() can be called before we have a kthread set up,
hence the need to check for its existence.