ilocks->irqsave locks
All ilocks are irqsave locks (the 'i' stands for interrupt).  This requires an
initialization of the spinlock.  Actually, all spinlocks should be initialized;
but you'll actually be called out on it with IRQ locks.  (The non-irqsave lock
initialization happens to be just 0).

Any new ilocks (from code from other plan9 devices) will need to have their
locks initialized.

Note that the locks inside apipes are not irqsave.  This might be an issue
later, if we try to push into a queue from IRQ context (but not from
RKMs/ktasks).