Helper to spawn kernel tasks (ktasks)
These are just kernel messages that might block.  Some subsystems launch
them, and they are meant to run for the life of the system.  We added a
name for debugging.

Also, since it's not clear how to schedule KMSGs or ktasks, we just send
them to the calling core, like with all KMSGs...