Kernel alarm tchains use locks
Previously, we left this to the tchain source to determine if it needed
locking or not.  The pcpu tchains were able to get by with IRQ
disabling.  This won't work if we want to be able to reasonably disable
an alarm, since we might want to disable from another core.