9ns uses rendezvous
It looks like all the existing tsleep() calls are just using it as a way
to sleep for a little while ("return0" sounds like it will never break
out from the cond() call, and also they were using a generic RV in the
proc).  So we just call udelay_sched() instead, which does the same
thing, without the mess.

Hopefully someone will come along that wants to use the timeouts, if for
nothing else than to get a decent test.  I'm a little surprised no one
uses them (like the TCP stack).

Also, if there's more 9ns code to bring in, the spatch script should
patch them up.