Improve debugging a bit and fix a simple bug
For some reason I needed to move code around a bit in iphash,
or I got very bogus hashes?

The added debugging let me get a bit further with the tcp connection bug.
The server in the tcp loopback test is forking and the child, when it writes
accept
is getting an error for no reason I can see.

Signed-off-by: Ronald G. Minnich <rminnich@google.com>