Devtabs initialize all methods
It looks like the style for struct dev is to fill in the entries, using
devwhatever() for ones the specific device doesn't provide.  This patch does so
for the hacked on power and config methods, which were implicitly 0 before.

I'll need this so I can add another op that will come after power and config.

The alternative is to initialize each member explicitly (.init = devinit, etc).
Its more of a question of how you handle your entries that the device does
nothing with: either 0 or a default.  If we went with 0, then the explicit
initialization is the way to go.  Since the style seems to be default
initializers, then we should just make sure everything has a default method.