Use the original plan 9 qio.c
Go back to the plan 9 version. With this change, nixtcploopback works.
Ping works.

But the ipconfig script now dies in the ethernet handler interrupt, need to
see why. It's an issue with calling qget in an interrupt handler.

Remove documentation/qio.c.txt, although why git decided this was
a rename we may never know.

Signed-off-by: Ronald G. Minnich <rminnich@google.com>