Fixes bug in errpop()
We weren't handling cur_index properly in errpop.

This one looked like a double-free of a chan in namec().  We entered the
second waserror() twice.  Once on a legit error, and a second time from
a failed nexterror().  (Failed due to errpop not advertising the correct
slot).

This would only come up in errors with an error in nested waserrors().