Fixes netif binding
We were setting the netfile's type to 0, which means we would never
detect/match any incoming packets.  This includes broadcast packets that
we sent ourselves (such as an arp, telling us that we are 10.0.2.15).

The root issue is atoi vs strtol.  atoi is supposed to be for base 10,
though NxM uses atoi for netifwrite.  Perhaps their strtol allows a base
prefix to override the base parameter.