Fixes netif binding
We were setting the netfile's type to 0, which means we would never
detect/match any incoming packets.  This includes broadcast packets that
we sent ourselves (such as an arp, telling us that we are 10.0.2.15).

The root issue is atoi vs strtol.  I ran into this issue a while ago,
and atoi should catch any further cases of this.