Queue limits use data length
Queues were using the allocated length of the queue's blocks when
judging the size of the queue.  Now they use the data length.  The
data length is the useful bytes in the queue.  The allocated length
includes extra padding and whatnot.  For instance, every block from the
igbe driver has an allocated size of 2KB, but most packets are much
smaller than this.

Prior to this change, the queue limit-check limits the amount of memory
consumed by the queue.  Now, the limit-check limits the amount of data
in the queue.  Since the network code / queue consumer is the one that
sets the limits, this makes more sense.  For instance, if network stack
code (like TCP) establishes queue limits (64K, 192K, etc), then you'd
think those limits should have something to do with actual network data,
not with whatever block allocation the driver happens to use.