Attempts to fix netifgen
If you bound #l0 to /net, then tried to ls /9/net/ether0 (without any
conversations in there), then we'd lock up.

The root of the problem goes back to ecff8a, when I tried fixing our
inability to stat /9/net/ether0/1/ifstats.  In the process, I broke
stating /9/net/ether0/{stats,ifstats}, which you'd only notice if you
didn't have any convs in ether0.

The root issue is that netif code can't tell the difference between a
QID with a netid of 0 and one with no netid at all.  I think I fixed it.
I left in my comments for the next time we need to debug this.