Add connection services (cs).
Add the pipe system call from NXM.

Add tests/common, an ndb file cs uses; add commands to makefile
to copy it when fill-kfs target is used.

cs opens a pipe and gets the fd, which is 6, but the plan 9 fd is 0.
cs opens #s/cs and wants to write the fd of the pipe to it, and
writes 6. #s barfs because that's not a valid fd -- the valid fd
is 0.

For now, srv will do the translation from akaros to plan 9 fd, but this
is getting increasingly nasty.

Clean up pipe a bit.

To test:
ash ifconfig
cs [-f /lib/ndb/common]
the optional argument should not be needed ;-)

Signed-off-by: Ronald G. Minnich <rminnich@google.com>