sys_fd2path (XCC)
Given an FD that corresponds to a 9ns chan, sys_fd2path() fills in u_buf
with up to len bytes of the chan's path.

Note this uses the path that was opened, not necessarily the actual
'value' of the chan.  For instance, if you open /net/tcp/clone, the path
is "/net/tcp/clone", but the chan value is for /net/tcp/N/ctl.  The Go
networking code expects this behavior.

Reinstall your kernel headers.