Get a little further with csquery
There's a bug in akaros. If you use gets() or read from
fd 0, for one byte, the system just locks up. Not sure
what the issue is.

But if you do this:
ash ifconfig
cs&
mount '#s/cs' /net
csquery /9/net/cs localhost

You'll see a full RPC with CS and you'll get an error back
(not sure what that is yet).

So we're closer. But failing reads on fd 0 needs fixin'.

Signed-off-by: Ronald G. Minnich <rminnich@google.com>