Add a simple ramfs from plan 9
Not quite working, probably due to Akaros / Plan 9 issues.

This is a ramfs in < 1000 lines. But, more importantly, it's
a roadmap on how to build a server. Obviously, this is much
better done in Go.

Usage:
ramfs&
mount '#s/ramfs' /mnt

And then you can in theory create files in /9/mnt, if it worked.
Debugging is hardwired on. Do we have the %F format yet for printing
9p packets?

Signed-off-by: Ronald G. Minnich <rminnich@google.com>