x86 serial input converts \r to \n
This is probably part of an ongoing disconnect at the console level
between the kernel and userspace.

The serial port reads a \r for a newline (at least in qemu and in
minicom a while back).  Userspace in general can't handle that.  csquery
couldn't, and busybox also has issues (try rm on
/9/bin/somefile-that-exists, you'll hang on the "really remove this?"
query).