Fixes kmalloc bug in mntwalk
It seems that any plan9 walk is going to have the problem.  We had a
similar issue with devwalk, and it looked the same.  (Canary clobber in
namec for e.elems, and the clobber was due to a previous (in memory, not
in time) kmalloc).

I also added some debugging stuff (single-core only) to help with next
time.