Kmalloc canary debugging
This was originally used to fix a kmalloc bug in mntwalk.

The issue that pops up with any 9ns code doing its own walks was messing
up with the wq size:

-	wq = kzmalloc(sizeof(struct walkqid) +
-		      (nname - 1) * sizeof(struct qid), 0);
+	wq = kzmalloc(sizeof(struct walkqid) + (nname) * sizeof(struct qid),
+				  KMALLOC_WAIT);

It seems that any plan9 walk is going to have the problem.  We had a
similar issue with devwalk, and it looked the same.  (Canary clobber in
namec for e.elems, and the clobber was due to a previous (in memory, not
in time) kmalloc).