Fix up of file open flags (XCC)
Plan9 and glibc differed in their open flags, as well as semantics on
things like CREATE.  Note there is no OCREAT in plan9, so we've got more
of a mashup of flags and syscalls.

I tried to get the intent of openmode() correct: reduce to the access
mode type, with EXEC => RDONLY, and ignore all the other flags.

devsrv was aborting for TRUNC, though I don't see why.  These fake
devices should be ignoring it anyways; everything is one big write from
the beginning, for all writes.

Reinstall your kernel headers, if you want access to the flags from
userspace.