Alarm to abort syscalls, used in ping
Will need to reorganize a lot of userspace's libraries.  Stuff like
nsec2tsc() doesn't need to be in a benchmarking library.  Same goes for
the alarm stuff; benchutil is more of a "everything I can't put in
parlib that we want by default."

Also, I cleaned up ping a little, notably the interrupted read code and
the spacing between sends (and now we get reasonable RTTs).