Adds some features to 'root' script
It'll still work as before for qemu users: ash root will ifconfig and
mount the same as before (now with a new srv name).

If you want to mount a UFS for one of the other machines that regularly
(but not always) run UFS, ifconfig first, then call for them by name:

$ ash ifconfig c89
$ ash root c75
$ ash root s141

etc.

Feel free to add your own entries to ifconfig and root.